Apply patch #89859 from JAn Wolter to fix Local Contrast tool settings conversion from/to XML for BQM workflow CCBUGS: 318726

+++ libs/dimg/filters/lc/localcontrastcontainer.cpp
@@ -60,14 +60,14 @@ LocalContrastContainer::~LocalContrastContainer()
 {
 }
 
-float LocalContrastContainer::getPower(int nstage) const
+double LocalContrastContainer::getPower(int nstage) const
 {
     float power = stage[nstage].power;
     power       = (float)(qPow(power / 100.0, 1.5) * 100.0);
     return power;
 }
 
-float LocalContrastContainer::getBlur(int nstage) const
+double LocalContrastContainer::getBlur(int nstage) const
 {
     return stage[nstage].blur;
 }
